rag-practise/file-loader/requirements.txt
langchain==0.3.11
langchain-core
langchain-openai==0.3.11
langchain-anthropic==0.3.11
langchain_community==0.3.11
tesseract
llama-index==0.12.15
llama-index-core==0.12.15
llama-index-readers-web

unstructured
unstructured-client==0.29.0
unstructured-inference==0.8.7
unstructured.pytesseract==0.3.13

pymilvus
pymilvus.model
milvus
langchain-deepseek
